First take on implementing a much more pythonic jax-toggle system.
Relies on creating classes with an exposed API. The problem is that JAX doesn't like acting on class methods so I am having to build around it.
